ki_pctcpu is unsigned, so we cast to long before subtracting.
Former-commit-id: a5160d19d5944ab2ec2d06d365fd972a04ac691c

+++ usr.bin/top/machine.c
@@ -3695,7 +3695,7 @@ name|a
 parameter_list|,
 name|b
 parameter_list|)
-value|do { \ 	long diff = (b)->ki_pctcpu - (a)->ki_pctcpu; \ 	if (diff != 0) \ 		return (diff> 0 ? 1 : -1); \ } while (0)
+value|do { \ 	long diff = (long)(b)->ki_pctcpu - (long)(a)->ki_pctcpu; \ 	if (diff != 0) \ 		return (diff> 0 ? 1 : -1); \ } while (0)
 end_define
 
 begin_define
